requirements.txt
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
google-cloud-aiplatform==1.36.4
anthropic==0.7.4
wordpress-api==1.0.0
psycopg2-binary==2.9.9
click==8.1.7
clickup-python==0.1.2
beautifulsoup4==4.12.2
python-wordpress-xmlrpc==2.3